+++ httprunner/testcase.py
@@ -265,7 +265,9 @@ class RequestWithOptionalArgs(object):
         self.__step_context.request.upload.update(file_info)
         return self
 
-    def teardown_hook(self, hook: Text, assign_var_name: Text = None) -> "RequestWithOptionalArgs":
+    def teardown_hook(
+        self, hook: Text, assign_var_name: Text = None
+    ) -> "RequestWithOptionalArgs":
         if assign_var_name:
             self.__step_context.teardown_hooks.append({assign_var_name: hook})
         else: